(I have failed at making this a one-thing commit))
Begun adding native in-engine documentation.
Fixed steering assistance (as well as other control configs) not applying.
Fixed the graph not drawing properly, and changed where/how it is loaded in-script.
Did a little more work on the in-engine VItaVehicle GUI.
Disabled VSync in order to display total FPS being reached (for performance measuring).